(test): report true per-test duration and de-duplicate variable build tests
The PHPUnit test reporter printed `durationSinceStart()` — wall-clock
elapsed since the run began — and labeled it as per-test duration. That
made the numbers climb monotonically across the run and flagged nearly
every test past the 15s mark as "slow", even pure CRUD tests that finish
in well under a second.

Capture an elapsed-since-start stamp when each test is prepared and
subtract it at finish, so the reported time and the "slow" warning are
the test's own duration.

Separately, the variable-in-build integration tests
(testProjectVariableInFunction, testProjectVariableInSite) lived in the
shared VariablesBase trait and therefore ran in both the console-client
and custom-server suites — 4 real Docker/npm builds where 2 suffice. The
contract is auth-side independent, so move them into
VariablesCustomServerTest only and keep the cheap CRUD coverage shared.
