+++ tests/extensions/TestHook.php
@@ -13,6 +13,7 @@ class TestHook implements Extension
 
     public function bootstrap(Configuration $configuration, Facade $facade, ParameterCollection $parameters): void
     {
+        $facade->registerSubscriber(new TestPreparedSubscriber());
         $facade->registerSubscriber(new TestFinishedSubscriber(self::MAX_SECONDS_ALLOWED));
         $facade->registerSubscriber(new RetrySubscriber());
     }
